+++ lib/core/services/ledgerservice.dart
@@ -8,7 +8,7 @@ class LedgerItemService {
   final CollectionReference companyCollection =
       Firestore.instance.collection('company');
 
-      Stream<List<LedgerItem>> get inactiveCustomerData {
+  Stream<List<LedgerItem>> get inactiveCustomerData {
     return companyCollection
         .document(this.uid)
         .collection('ledger')
@@ -18,7 +18,7 @@ class LedgerItemService {
         .map(_ledgerItemData);
   }
 
-        Stream<List<LedgerItem>> get accountsReceivableData {
+  Stream<List<LedgerItem>> get accountsReceivableData {
     return companyCollection
         .document(this.uid)
         .collection('ledger')
@@ -27,7 +27,7 @@ class LedgerItemService {
         .map(_ledgerItemData);
   }
 
-        Stream<List<LedgerItem>> get accountsPayablesData {
+  Stream<List<LedgerItem>> get accountsPayablesData {
     return companyCollection
         .document(this.uid)
         .collection('ledger')
@@ -51,8 +51,8 @@ class LedgerItemService {
         name: doc.data['name'].toString() ?? '',
         masterId: doc.data['master_id'].toString() ?? '',
         currencyName: doc.data['currencyname'].toString() ?? '',
-        openingBalance: doc.data['opening_balance'].toInt() ?? 0,
-        closingBalance: doc.data['closing_balance'].toInt() ?? 0,
+        openingBalance: doc.data['opening_balance'].toDouble() ?? 0,
+        closingBalance: doc.data['closing_balance'].toDouble() ?? 0,
         parentid: doc.data['parentcode'].toString() ?? '',
         contact: doc.data['contact'].toString() ?? '',
         state: doc.data['state'].toString() ?? '',
@@ -61,26 +61,32 @@ class LedgerItemService {
         guid: doc.data['guid'].toString() ?? '',
         gst: doc.data['gst'].toString() ?? '',
         lastPaymentDate: doc.data['restat_last_payment_date']?.toDate() ?? null,
-          lastPurchaseDate: doc.data['restat_last_purchase_date']?.toDate() ?? null,
-          lastReceiptDate: doc.data['restat_last_receipt_date']?.toDate() ?? null,
-          lastSalesDate: doc.data['restat_last_sales_date']?.toDate() ?? null,
-          meanPayment: doc.data['restat_mean_payment'].toString()?? '',
-          meanPurchase: doc.data['restat_mean_purchase'].toString() ?? '',
-          meanReceipt: doc.data['restat_mean_receipt'].toString() ?? '',
-          meanSales: doc.data['restat_mean_sales'].toString() ?? '',
-          partyGuid: doc.data['restat_party_guid'].toString() ?? '',
-          totalPayables: doc.data['restat_total_payables'].toString() ?? '',
-          totalPayment: doc.data['restat_total_payment'].toString() ?? '',
-          totalSales: doc.data['restat_total_sales'].toString() ?? '',
-          totalPurchase: doc.data['restat_total_purchase'].toString() ?? '',
-          totalReceipt: doc.data['restat_total_receipt'].toString() ?? '',
-          totalReceivables: doc.data['restat_total_receivables'].toString() ?? '',
-          primaryGroupType: doc.data['restat_primary_group_type'].toString() ?? '',
-          restatCompanyCode: doc.data['restat_company_code'].toString() ?? '',
-          numPaymentVouchers: doc.data['restat_num_payment_vouchers'].toString() ?? '',
-          numPurchaseVouchers: doc.data['restat_num_purchase_vouchers'].toString() ?? '',
-          numReceiptVouchers: doc.data['restat_num_receipt_vouchers'].toString() ?? '',
-          numSalesVouchers: doc.data['restat_num_sales_vouchers'].toString() ?? '',
+        lastPurchaseDate:
+            doc.data['restat_last_purchase_date']?.toDate() ?? null,
+        lastReceiptDate: doc.data['restat_last_receipt_date']?.toDate() ?? null,
+        lastSalesDate: doc.data['restat_last_sales_date']?.toDate() ?? null,
+        meanPayment: doc.data['restat_mean_payment'].toString() ?? '',
+        meanPurchase: doc.data['restat_mean_purchase'].toString() ?? '',
+        meanReceipt: doc.data['restat_mean_receipt'].toString() ?? '',
+        meanSales: doc.data['restat_mean_sales'].toString() ?? '',
+        partyGuid: doc.data['restat_party_guid'].toString() ?? '',
+        totalPayables: doc.data['restat_total_payables'].toString() ?? '',
+        totalPayment: doc.data['restat_total_payment'].toString() ?? '',
+        totalSales: doc.data['restat_total_sales'].toString() ?? '',
+        totalPurchase: doc.data['restat_total_purchase'].toString() ?? '',
+        totalReceipt: doc.data['restat_total_receipt'].toString() ?? '',
+        totalReceivables: doc.data['restat_total_receivables'].toString() ?? '',
+        primaryGroupType:
+            doc.data['restat_primary_group_type'].toString() ?? '',
+        restatCompanyCode: doc.data['restat_company_code'].toString() ?? '',
+        numPaymentVouchers:
+            doc.data['restat_num_payment_vouchers'].toString() ?? '',
+        numPurchaseVouchers:
+            doc.data['restat_num_purchase_vouchers'].toString() ?? '',
+        numReceiptVouchers:
+            doc.data['restat_num_receipt_vouchers'].toString() ?? '',
+        numSalesVouchers:
+            doc.data['restat_num_sales_vouchers'].toString() ?? '',
       );
     }).toList();
   }
