+++ lib/ui/shared/debitcredit.dart
@@ -1,23 +1,13 @@
-
 import 'package:tassist/core/services/string_format.dart';
 
-
- debitCredit(int inputNumber) {
-  
+debitCredit(double inputNumber) {
   if (inputNumber != null) {
-  
-  if (inputNumber < 0) {
-      return '${formatIndianCurrency(inputNumber.toString().substring(1, inputNumber.toString().length))} Dr.'; 
-     
-  }
-  else {
-    return '${formatIndianCurrency(inputNumber.toString())} Cr.';
-   
-  }
-  }
-
-  else {
+    if (inputNumber < 0) {
+      return '${formatIndianCurrency(inputNumber.toString().substring(1, inputNumber.toString().length))} Dr.';
+    } else {
+      return '${formatIndianCurrency(inputNumber.toString())} Cr.';
+    }
+  } else {
     return inputNumber;
   }
-
 }
